+++ code/java-netty/src/main/java/com/zync/chat/protocol/MessageCodecSharable.java
@@ -37,8 +37,8 @@ public class MessageCodecSharable extends MessageToMessageCodec<ByteBuf, Message
         out.writeByte(algorithm.getSerializerAlgorithm());
         // 4. 1 字节的指令类型
         out.writeByte(msg.getMessageType());
-        // 5. 4 个字节
-        out.writeInt(msg.getSequenceId());
+        // 5. 8 个字节
+        out.writeLong(msg.getSequenceId());
         // 无意义，对齐填充
         out.writeByte(0xff);
         // 6. 获取内容的字节数组
@@ -56,7 +56,7 @@ public class MessageCodecSharable extends MessageToMessageCodec<ByteBuf, Message
         byte version = in.readByte();
         byte serializerAlgorithm = in.readByte();
         byte messageType = in.readByte();
-        int sequenceId = in.readInt();
+        long sequenceId = in.readLong();
         in.readByte();
         int length = in.readInt();
         byte[] bytes = new byte[length];
